+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-05-15 12:38:21
-// Commit: feat(test): add bi-directional event bus handler for test
+// Updated at: 2026-05-15 12:58:34
+// Commit: feat(test): implement shimmer loading skeleton in test
 
 class testComponentState {
-  final String commitRef = '1182';
-  final String lastMessage = 'feat(test): add bi-directional event bus handler for test';
-  final DateTime timestamp = DateTime.parse('2026-05-15T12:38:21+05:00');
+  final String commitRef = '1183';
+  final String lastMessage = 'feat(test): implement shimmer loading skeleton in test';
+  final DateTime timestamp = DateTime.parse('2026-05-15T12:58:34+05:00');
 }